refactor(ts): FooterToolbar
