--- source/components/FooterToolbar/index.js
@@ -1,4 +1,0 @@
-import './style/index.less';
-
-export default from './FooterToolbar';
-
