+++ source/components/FooterToolbar/index.tsx
@@ -1,0 +1,4 @@
+import FooterToolbar from './FooterToolbar';
+import './style/index.less';
+
+export default FooterToolbar;
